refactor(common/HttpManagerApiRepository): rename & move in its own file
