+++ libs/ledger-live-common/src/device/hooks/useLatestFirmware.ts
@@ -1,13 +1,13 @@
 import { getProviderId } from "../../manager/index";
 import { getEnv } from "@ledgerhq/live-env";
 import { version } from "../../../package.json";
-import { ManagerApiHttpRepository } from "../../device-core/repositories/ManagerApiRepository";
+import { HttpManagerApiRepository } from "../../device-core/repositories/HttpManagerApiRepository";
 import { useGetLatestFirmware } from "../../device-react/useGetLatestFirmware";
 import { DeviceInfoEntity } from "../../device-core/entities/DeviceInfoEntity";
 
 export function useLatestFirmware(
   deviceInfo?: DeviceInfoEntity | null,
-  managerApiRepository: ManagerApiHttpRepository = new ManagerApiHttpRepository(
+  managerApiRepository: HttpManagerApiRepository = new HttpManagerApiRepository(
     getEnv("MANAGER_API_BASE"),
     version,
   ),
