+++ libs/ledger-live-common/src/device/use-cases/managerApi/fetchMcusUseCase.ts
@@ -1,9 +1,9 @@
 import { version } from "../../../../package.json";
 import { getEnv } from "@ledgerhq/live-env";
-import { ManagerApiHttpRepository } from "../../../device-core/repositories/ManagerApiRepository";
+import { HttpManagerApiRepository } from "../../../device-core/repositories/HttpManagerApiRepository";
 
 export default function fetchMcusUseCase(
-  managerApiRepository = new ManagerApiHttpRepository(getEnv("MANAGER_API_BASE"), version),
+  managerApiRepository = new HttpManagerApiRepository(getEnv("MANAGER_API_BASE"), version),
 ) {
   return managerApiRepository.fetchMcus();
 }
